% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_name.R
\name{author_name}
\alias{author_name}
\title{Pull first name(s) and last name for a given list of orcid}
\usage{
author_name(
  data,
  first_name = "first_name",
  last_name = "last_name",
  initial = T,
  position = "right"
)
}
\arguments{
\item{data}{datafame containing a vectors of the author name (split into first_name and last_name)}

\item{first_name}{Column name of vector containing the first and middle name(s) (default = "first_name")}

\item{last_name}{Column name of vector containing the last or family name (default = "last_name")}

\item{initial}{Should the first / middle name(s) be converted to initial (default = TRUE)}

\item{position}{initial to "left" or "right" of last name (default = "right")}

\item{initial_max}{Maximum number of digits (default = 3)}
}
\value{
Vector of the combined name composing the full author name in the requested format.
}
\description{
Pull and format first name(s) and last name for a given list of orcid
}
