% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_01.R
\name{collapse01}
\alias{collapse01}
\title{Collapse a group of binary variables (e.g. those generated by checkbox variables in REDCap) into summary columns.}
\usage{
collapse01(df, cols, prefix = NULL, suffix = NULL, output = c("list",
  "yesno", "n"), remove = TRUE, binary0 = c(0, "0", "No", "no"),
  binary1 = c(1, "1", "Yes", "yes"))
}
\arguments{
\item{df}{Dataframe.}

\item{cols}{List of columnn names of binary variables desired to be collapsed.}

\item{prefix}{String to add to the start of all summary columns names.}

\item{suffix}{String to add to the end of all summary columns names.}

\item{output}{List of desired outputs: yesno (any values selected), n (number of values selected) and list (list of column names of all values selected).}

\item{remove}{Logical value to remove columns supplied to the "cols" parameter}

\item{binary0}{List of all values corresponding to "0" (No) in the binary variable (default = 0 or no)}

\item{binary1}{List of all values corresponding to "1" (Yes) in the binary variable (default = 1 or yes).}
}
\value{
Dataframe with up to 3 additional columns: yesno (any values selected), n (number of values selected) and list (list of column names of all values selected).
}
\description{
Collapse a group of binary variables (e.g. those generated by checkbox variables in REDCap) into summary columns.
}
