% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_manage.R
\name{dag_manage}
\alias{dag_manage}
\title{Generate REDCap summary data.}
\usage{
dag_manage(
  redcap_project_uri,
  redcap_project_token,
  import = NULL,
  remove = NULL
)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{import}{A list of new DAGs to import into the project (Default = NULL).}

\item{remove}{A list of current DAGs to delete in the project (Default = NULL).}
}
\value{
A dataframe of DAGs specifiying those which are new, deleted, or unchanged (-).
}
\description{
Used to generate high quality summary data for REDCap projects at overall, and DAG-specific level.
}
