% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_author.R
\docType{data}
\name{data_author}
\alias{data_author}
\title{Example dataset for authorship.}
\format{A data frame with 100 rows and 3 variables:
\describe{
  \item{name}{randomly generated last name and initial}
  \item{hospital}{example list of hospitals}
  \item{country}{example list of countries}
  ...
}}
\usage{
data(data_author)
}
\description{
A dataset containing randomly generated names for demonstrating how to generate an authorship list.
}
\keyword{datasets}
