% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_body.R
\name{email_body}
\alias{email_body}
\title{Split a tibble/dataframe by "group" variable, then save HTML file and/or code (e.g. for email body in send_email function)}
\usage{
email_body(
  data,
  group = "group",
  rmd_file,
  html_output = c("code", "file"),
  subfolder = here::here("folder_html"),
  file_prefix = "",
  file_suffix = ""
)
}
\arguments{
\item{data}{Dataframe with at least 1 column - corresponding to the "group".}

\item{group}{Grouping variable (must be unique values) who will recieve unique email.}

\item{rmd_file}{String of the path of the RMD file with an HTML output (email body) within working directory.}

\item{html_output}{List of desired html outputs with "code" (html code) and "file" options (html file path). (Default = both "code" and "file")}

\item{subfolder}{Folder within working directory where HTML file will be stored - only required if file %in% html_output. (Default = "folder_html")}

\item{file_prefix}{String to be prefixed to "group" when naming HTML file - only required if file %in% html_output.}

\item{file_suffix}{String to be suffixed to "group" when naming HTML file - only required if file %in% html_output.}
}
\value{
Original dataframe with up to 2 additional columns: html code ("code") and/or html file path ("file") depending on html_output.
}
\description{
Split a tibble/dataframe by "group" variable, then save HTML file and/or code (e.g. for email body in send_email function)
}
