% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_field.R
\name{email_field}
\alias{email_field}
\title{Prepare df_email in a standard format for email_send function.}
\usage{
email_field(
  df_email,
  group = "data_access_group",
  subject,
  recipient_main = NULL,
  recipient_cc = NULL,
  recipient_bcc = NULL
)
}
\arguments{
\item{df_email}{Dataframe with at least 2 columns - corresponding to a "group" and and "email" (e.g. output from user_summarise). Email addresses must be a string separated by a comma or semi-colon (not a list variable).}

\item{group}{Grouping variable (must be unique values) who will recieve unique email.}

\item{subject}{String corresponding to the subject line of the emails - subjects can be made group specific using square brachets "[]" around a column name within the dataset.}

\item{recipient_main}{String corresponding to column name containing the main intended email recipients within the group.}

\item{recipient_cc}{String corresponding to column name containing email recipients within the group intended to be cc'd.}

\item{recipient_bcc}{String corresponding to column name containing email recipients within the group intended to be bcc'd.}
}
\value{
Dataframe with 5 mandatory columns: group, recipients (recipient_main, recipient_cc & recipient_bcc), and subject.
}
\description{
Prepare df_email in a standard format for email_send function.
}
