% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_send.R
\name{email_send}
\alias{email_send}
\title{Send group-specific emails and attatchments via gmailR.}
\usage{
email_send(
  df_email,
  sender,
  body = "code",
  attach = NULL,
  zip = F,
  draft = TRUE
)
}
\arguments{
\item{df_email}{Dataframe containing a minimum of the output from email_format function and a column corresponding to the email body pathway.}

\item{sender}{Email address from the sender (should match the details specified using gmailr::gm_auth_configure()).}

\item{body}{String of column names of the HTML code to attach as the email body (default = "code"). Note html files cannot be sent at present using gmailR.}

\item{attach}{List of column names of files to attach. Each column must contain a full path for each file for each group.}

\item{zip}{Logical value to send attachments as a ZIP folder (default = FALSE). Requires an attachment via "attach".}

\item{draft}{Logical value to create as a draft email (default: TRUE) or send immediately (FALSE)}
}
\value{
Vector of paths for each group-specific HTML file (append to df_email to use in send_email function)
}
\description{
Send group-specific emails and attatchments via gmailR (must have previously set up GmailR on R instance)
}
