% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data_dict}
\alias{example_data_dict}
\title{Example dataset of surgical patients}
\format{A data frame with 20 rows and 11 variables.}
\usage{
data(example_data_dict)
}
\description{
Example dataset to be used within the \href{https://github.com/kamclean/collaborator/blob/master/vignettes/vignette_data_dict.md}{Generating a Simple, Easily-Shareable Data Dictionary} vignette (all data randomly generated).
}
\keyword{data}
