% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_df_user}
\alias{example_df_user}
\title{Example dataset of REDCap users}
\format{
A data frame with 34 rows and 2 variables.
}
\usage{
data(example_df_user)
}
\description{
Example dataset to be used within the \href{https://github.com/kamclean/collaborator/blob/master/vignettes/vignette_user_1_explore.md}{Redcap User Management: 1. Explore Current Users} vignette (all data randomly generated).
}
\keyword{data}
