% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_report_author}
\alias{example_report_author}
\title{Example dataset of authors}
\format{A data frame with 100 rows and 3 variables.}
\usage{
data(example_report_author)
}
\description{
Example dataset to be used within the \href{https://github.com/kamclean/collaborator/blob/master/vignettes/vignette_authors.md}{Generating Authorship Lists} vignette (all data randomly generated).
}
\keyword{data}
