% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.R
\name{get_drive_email}
\alias{get_drive_email}
\title{pull drive email from environment}
\usage{
get_drive_email(drive_email = NULL)
}
\arguments{
\item{drive_email}{optional str, email address, if not provided will use envronment}
}
\value{
string, the email to use, either what was sent or from the config
}
\description{
convenience method for replacing an empty drive, called by drive_setup
}
