% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group2csv.R
\name{group2csv}
\alias{group2csv}
\title{Split a tibble/dataframe into CSV by "group" variable}
\usage{
group2csv(data, group, subfolder = "folder_csv", file_prefix = "",
  file_suffix = "")
}
\arguments{
\item{data}{Dataframe with at least 1 column - corresponding to a "group".}

\item{group}{Grouping variable (must be unique values) who will recieve unique email.}

\item{subfolder}{Folder within working directory (e.g. string entered into here::here()) where CSV will be stored. Default = "folder_csv".}

\item{file_prefix}{String to be prefixed to "group" when naming CSV file.}

\item{file_suffix}{String to be suffixed to "group" when naming CSV file.}
}
\value{
Dataframe of group AND csv file path ("file").
}
\description{
Split a tibble/dataframe by "group" variable, then save in a subfolder as CSV.
}
