% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_valid.R
\name{orcid_valid}
\alias{orcid_valid}
\title{Validate vector of ORCID}
\usage{
orcid_valid(data, orcid = "orcid", reason = FALSE, na.rm = FALSE)
}
\arguments{
\item{data}{datafame containing a vector of ORCID}

\item{orcid}{Column name of vector containing ORCID (default = "orcid")}

\item{reason}{Logical value to determine whether output should include reasons for validity (default = FALSE) or vector of ORCID (TRUE).}

\item{na.rm}{Remove NA (invalid ORCID) from output}
}
\value{
Vector of orcid (reason = FALSE) or tibble with columns specifying the validation checks failed by the ORCID ("check_" columns)
}
\description{
Validate vector of ORCID based number of digits / format / checksum.
}
