% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_pull_name.R
\name{pull_orcid_name}
\alias{pull_orcid_name}
\title{Pull first name(s) and last name for a given list of orcid}
\usage{
pull_orcid_name(list_orcid, initials = TRUE, full_name = TRUE)
}
\arguments{
\item{list_orcid}{List of orcid ids (XXXX-XXXX-XXXX-XXXX format)}

\item{initials}{Should a column containing names with up to 3 initals be included? (default = TRUE)}

\item{full_name}{Should a column containing the full name be included? (default = TRUE)}
}
\value{
Dataframe with 3 mandatory columns: orcid, first names (fn_orcid) and last name (ln_orcid)
}
\description{

}
