% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{read_data_specification}
\alias{read_data_specification}
\title{read a spec sheet of columns descibing data to be validated}
\usage{
read_data_specification(csv_file)
}
\arguments{
\item{csv_file}{a csv with column col_name and col_type}
}
\value{
data frame that meets criteria
}
\description{
read a spec sheet of columns descibing data to be validated
}
