% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_compare.R
\name{redcap_compare}
\alias{redcap_compare}
\title{Compare multiple REDCap projects to determine discrepancies in structure or user rights.}
\usage{
redcap_compare(redcap_project_uri, redcap_token_list, comparison)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance (must all be on same REDCap instance)}

\item{redcap_token_list}{List of API (Application Programming Interface) for the REDCap projects.}

\item{comparison}{What should be compared - project structure ("metadata") or user roles ("role").}
}
\value{
Nested tibble of the full comparison across projects ("full") and the specific discrepancies ("discrepancies").
}
\description{
Used to compare multiple REDCap projects to determine discrepancies in structure or user rights.
}
