% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_format_repeat.R
\name{redcap_format_repeat}
\alias{redcap_format_repeat}
\title{Change structure of repeating data}
\usage{
redcap_format_repeat(data, format = "long")
}
\arguments{
\item{data}{Output from redcap_data$data}

\item{format}{The format the repeating instrument data should be provided in. Options include "long" (default), "wide" (each instance a separate column), or "list" (nested instances).}
}
\value{
Dataframe
}
\description{
Change structure of repeating data from redcap_data from long to either list or wide.
}
