% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_label.R
\name{redcap_label}
\alias{redcap_label}
\title{Export REDCap dataset and label using metadata}
\usage{
redcap_label(
  data = NULL,
  metadata = NULL,
  redcap_project_uri = NULL,
  redcap_project_token = NULL,
  use_ssl = TRUE,
  column_name = "raw",
  column_attr = NULL,
  checkbox_value = "label"
)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{use_ssl}{Logical value whether verify the peer's SSL certificate should be evaluated (default=TRUE)}

\item{column_name}{Determine if output column names should be unchanged from the REDCap record export ("raw") or labelled ("label"). Default = "raw".}

\item{column_attr}{Determine if a labelled attribute should be applied and whether this should be the original ("raw") or labelled ("label") name. Default = NULL.}

\item{checkbox_value}{Determine if output checkbox variables should be unchanged from the REDCap record export ("raw") or labelled ("label"). Default = "raw".}
}
\value{
Three nested tibbles: (1) "exported": REDcap record export (unchanged) (2) labelled": REDcap record export with variables classified and columns labelled as specified via column_name and column_attr (3) "metadata": Cleaned metadata file for the REDCap dataset.
}
\description{
Export the REDCap dataset, and use the metadata to classify the variables and label the columns.
}
