% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_log.R
\name{redcap_log}
\alias{redcap_log}
\title{Export the redcap log}
\usage{
redcap_log(
  redcap_project_uri,
  redcap_project_token,
  date_start = NULL,
  date_end = NULL,
  item = NULL
)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{date_start}{Limit to the start date extracted in the format of YYYY-MM-DD (default = NULL)}

\item{date_end}{Limit to the end date extracted in the format of YYYY-MM-DD (default = NULL)}

\item{item}{A list of all types of events wanted to be exported (default = NULL aka all). These can include "record", "user", "page_view","lock_record", "manage", "record_add", "record_edit", "record_delete", "export".}
}
\value{
Logging record of the specified events
}
\description{
Export the redcap log to gain insight into redcap changes over time
}
