% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_metadata.R
\name{redcap_metadata}
\alias{redcap_metadata}
\title{Export REDCap metadata (with individual checkbox variables if present) and variable class in R.}
\usage{
redcap_metadata(redcap_project_uri, redcap_project_token, descriptive = FALSE)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{descriptive}{Logical value whether to include descriptive fields within the dataset (default = FALSE)}
}
\value{
Tibble of REDCap project metadata (with individual checkbox variables if present) and variable class in R.
}
\description{
Used to generate high quality summary data for REDCap projects at overall, and DAG-specific level.
}
