% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_sim.R
\name{redcap_sim}
\alias{redcap_sim}
\title{Simulate data from REDCap metadata}
\usage{
redcap_sim(
  redcap_project_uri,
  redcap_project_token,
  nrecords = 100,
  propmiss = 0.1,
  sampledist = "normal"
)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{nrecords}{Numeric value of the number of unique records to simulate (default = 100). If there are repeating instruments, there will be more rows than this.}

\item{propmiss}{Proportion of data missing within each variable in the simulated data (only record_id is excempt)}

\item{sampledist}{A string specifying the data distribution for all numeric/date/datetime variables. Either "normal", "lskew", "rskew", "equal". (default = equal)}
}
\value{
Tibble of simulated REDCap data
}
\description{
Used to generate simulated data from REDCap metadata, incorporating repeating instruments, validation and branching logic (where present).
}
