% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_sum.R
\name{redcap_sum}
\alias{redcap_sum}
\title{Generate REDCap summary data.}
\usage{
redcap_sum(redcap_project_uri, redcap_project_token, centre_sum = T,
  n_top_dag = 10, var_exclude = "", var_complete = "",
  user_exclude = "", record_exclude = "", dag_exclude = "",
  record_dag_field = "redcap_data_access_group",
  user_dag_field = "data_access_group")
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{centre_sum}{Logical input (TRUE/FALSE) to determine whether data access group-level summaries will be produced (Default = TRUE).}

\item{n_top_dag}{When centre_sum = TRUE, defines output of the number of centres with the most records uploaded (default is top 10).}

\item{var_exclude}{Vector of names of variables that are desired to be excluded from assessment of data completness (any NA value will be counted as incomplete).}

\item{var_complete}{Vector of names of variables that are desired to be specifically used to assess data completness (alternate method from using "var_exclude").}

\item{user_exclude}{Vector of redcap usernames that are desired to be excluded from the user count (note all users not assigned to a DAG will automatically be excluded).}

\item{record_exclude}{Vector of redcap record_id that are desired to be excluded from the record count.}

\item{dag_exclude}{Vector of redcap data access group names that are desired to be excluded from the record count.}

\item{record_dag_field}{Name of the data access group variable in the redcap record API output (default is: redcap_data_access_group).}

\item{user_dag_field}{Name of the data access group variable in the redcap user API output (default is: data_access_group).}
}
\value{
Nested dataframes of (i) overall summary statistics for the project ("sum_overall") (ii). DAG-specific summary statistics for the project ("dag_all") (iii). DAGs with no data uploaded, but users assigned ("dag_nodata") (iv). DAGs with <100% completeness ("dag_incom") (v). The top n recruiting centres ("dag_top_n").
}
\description{
Used to generate high quality summary data for REDCap projects at overall, and DAG-specific level.
}
