% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_auth.R
\name{report_auth}
\alias{report_auth}
\title{Generate a formatted authorship list.}
\usage{
report_auth(df, group = NULL, subdivision = NULL, name_sep = ",",
  group_brachet = "()", group_sep = ";")
}
\arguments{
\item{df}{Dataframe of at least 2 columns: "name".}

\item{group}{Column name of a variable in the dataframe by which to group authorship (the default is NULL).}

\item{subdivision}{Column name of an additional variable in the dataframe by which to subdivide authorship (the default is NULL).}

\item{name_sep}{Character(s) which will separate names within the group (the default is ",").}

\item{group_brachet}{Character(s) bracheting the group (the default is "()").}

\item{group_sep}{Character(s) which will separate the groups (the default is ";").}
}
\value{
Generates a text file ("auth_out.txt").
}
\description{
Used to generate a formatted authorship list for all users by group (the group the user belongs to e.g. the centre which they participated at). Optional subdivisions can be created to stratify users and groups. This could be a role (e.g. collaborator and validator) or region/country.
}
