% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file_same.R
\name{upload_file_same}
\alias{upload_file_same}
\title{Upload the same file to all records}
\usage{
upload_file_same(redcap_project_uri, redcap_project_token, record_id,
  var_upload, name_file)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{record_id}{Vector of record ids which the file should be uploaded to.}

\item{var_upload}{The field name of the upload location}

\item{name_file}{The location of the file to be uploaded to all records in "record_id".}
}
\value{
Tibble of all records which the file could not be uploaded to.
}
\description{
Use to upload the same file to all records in record_id
}
