% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file_unique.R
\name{upload_file_unique}
\alias{upload_file_unique}
\title{Upload the same file to all records}
\usage{
upload_file_unique(redcap_project_uri, redcap_project_token, record_id,
  var_upload, name_folder, name_file_suffix)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{record_id}{Vector of record ids which the file should be uploaded to.}

\item{var_upload}{The field name of the upload location}

\item{name_folder}{The location of the folder with files to be uploaded to all records in "record_id".}

\item{name_file_suffix}{Each file must start with the appropriate record_id and end with a file extension (".csv", ".txt", etc). Further characters can be incorported if desired before the file extention (e.g. "_result.txt")}
}
\value{
Tibble of all records which the file could not be uploaded to
}
\description{
Use to upload the same file to all records in record_id
}
