% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_record_new.R
\name{upload_record_new}
\alias{upload_record_new}
\title{Upload new records to REDCap project}
\usage{
upload_record_new(redcap_project_uri, redcap_project_token, df_new)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{df_new}{Dataframe of record_id (mandatory), and other columns (optional) that match the names / required formats of the original dataset. Columns not specified will be uploaded as blank.}
}
\value{
REDCap project updated to include new record_id.
}
\description{
Use to upload new records to REDCap project
}
