% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_assign.R
\name{user_assign}
\alias{user_assign}
\title{Used to assign users to data access groups on a redcap project.}
\usage{
user_assign(redcap_project_uri, redcap_project_token, users.df, role)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{users.df}{Dataframe containing at least 2 columns (username, data_access_group)}

\item{role}{Username of user with the desired user rights}
}
\value{
None (user rights uploaded directly into REDCap - user acccounts are still required to be entered manually)
}
\description{
Used to assign users to data access groups on a redcap project with the same user rights as a current user (role). This can add new users or edit existing ones.
}
\details{
# Sources of errors: (i) DAG has not been added on REDCap (this must be done prior to users being assigned). (ii) Username is not in an acceptable format (e.g. contains spaces, unavaliable characters, etc)
}
