% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_import.R
\name{user_import}
\alias{user_import}
\title{Generate a csv file that can be used to upload new user accounts to REDCap directly (via control centre)}
\usage{
user_import(df, username, first_name, last_name, email, institution = "",
  sponser = "", expiration = "", comments = "")
}
\arguments{
\item{df}{Dataframe of at least 4 mandatory columns (corresponding to: username, first name, last name, and email address) and 4 optional columns (corresponding to: institution, sponser, expiration, comments). All optional columns will be blank unless otherwise specified.}

\item{username}{Mandatory variable (corresponds to "Username").}

\item{first_name}{Mandatory variable (corresponds to "First name").}

\item{last_name}{Mandatory variable (corresponds to "Last name").}

\item{email}{Mandatory variable (corresponds to "Email address").}

\item{institution}{Optional variable which is recommended to record the data_access_group / centre of the user (corresponds to "Institution ID").}

\item{sponser}{Optional variable (corresponds to "Sponsor username").}

\item{expiration}{Optional variable (corresponds to "Expiration"). Must be in YYYY-MM-DD HH:MM or MM/DD/YYYY HH:MM format}

\item{comments}{Optional variable (corresponds to "Comments").}
}
\value{
Returns a dataframe and csv file for upload ("user.import_[date generated].csv").
}
\description{
Us
}
