% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_manage.R
\name{user_manage}
\alias{user_manage}
\title{Used to manage REDCap project users}
\usage{
user_manage(
  redcap_project_uri,
  redcap_project_token,
  users = NULL,
  role = NULL,
  dag = NULL,
  remove = FALSE
)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{users}{Vector of usernames or a dataframe containing at least 1 column ("username"). Further columns specifying individual patients to be removed ("remove") or assigned to roles ("role") or data access groups ("dag") can be added.}

\item{role}{String of a single role ID, role name or username of user with the desired user rights to be applied to ALL users specified in "users"  (use a column in users if wanting to be different for each user). Must specify "none" if no assignment of role.}

\item{dag}{String of a unique DAG to ALL users specified in "users" will be assigned to (use a column in users if wanting to be different for each user). Must specify "none" if no assignment of DAG.}

\item{remove}{Logical value indicating if ALL users specified in "users" are to be removed  (use a column in users if wanting to be different for each user). Default is FALSE (no users to be removed).}
}
\value{
Nested tibbles of the outcome (1) "correct" users with the correct allocation specified (2) "error" users with an allocation unable to be completed. User acccounts are still required to be entered manually.
}
\description{
Used to manage users - whether to change users present, or their specific roles and data access groups on a redcap project.
}
