% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_new.R
\name{user_new}
\alias{user_new}
\title{Generates a dataframe of all new users}
\usage{
user_new(redcap_project_uri, redcap_project_token, user_all.df, users_exception)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{user_all.df}{Dataframe of all users containing at least 1 column ("username")}

\item{users_exception}{= Vector of any usernames to be excluded.}
}
\value{
Original dataframe containing only usernames not yet allocated to the REDCap project
}
\description{
Used to compare a dataframe of all usernames to those currently allocated on REDCap to determine any new users requiring assigned
}
