% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_role.R
\name{user_role}
\alias{user_role}
\title{Identifies unique REDCap user roles}
\usage{
user_role(
  redcap_project_uri,
  redcap_project_token,
  user_exclude = NULL,
  remove_id = T
)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{user_exclude}{Vector of usernames to be excluded (e.g. those with unique rights). Default is none (e.g. "").}
}
\value{
Dataframe of all users by unique role
}
\description{
Used to count the number of unique roles (e.g. unique combinations of user rights) on the REDCap project. Note: this replaces the function of roles on the user rights page of the REDCap.
}
