% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_roles.R
\name{user_roles}
\alias{user_roles}
\title{Assign named roles to REDCap users}
\usage{
user_roles(redcap_project_uri, redcap_project_token, role_users_example)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{role_users_example}{Dataframe with 2 columns: role (specifiying the name of the role), and username (the username with the desired rights to be associated with the role).}
}
\value{
Dataframe of REDCap project users with an additional "role" column.
}
\description{
Used to assign a role name to current redcap users based on the example users given in role_users_example. Note: The number of roles should match the output from user_roles_n.
}
