% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_role.R
\name{user_roles_n}
\alias{user_roles_n}
\title{Identifies unique REDCap user roles}
\usage{
user_roles_n(redcap_project_uri, redcap_project_token,
  users_ignore = NULL, use_ssl = TRUE)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{users_ignore}{Vector of usernames to be excluded (e.g. those with unique rights). Default is none (e.g. "").}
}
\value{
Nested dataframe of (i) Dataframe of all users numbered by unique role. (ii) Dataframe of each role with an example user with those user rights.
}
\description{
Used to count the number of unique roles (e.g. unique combinations of user rights) on the REDCap project. Note: this replaces the function of roles on the user rights page of the REDCap.
}
