% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validate_from_file}
\alias{validate_from_file}
\title{create validation results from a yaml file}
\usage{
validate_from_file(data_df, file)
}
\arguments{
\item{data_df}{data frame of data}

\item{file}{yaml formatted file with rules for the validate package}
}
\value{
the outupt from confront function
}
\description{
create validation results from a yaml file
}
