% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validation_report}
\alias{validation_report}
\title{create a report of validation results}
\usage{
validation_report(data.df, validation_file)
}
\arguments{
\item{data.df}{data frame to validate}

\item{validation_file}{validations rules}
}
\value{
string of validation results, empty string if no errors
}
\description{
create a report of validation results
}
