% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_name.R
\name{orcid_name}
\alias{orcid_name}
\title{Pull first name(s) and last name for a given list of orcid}
\usage{
orcid_name(data, orcid = "orcid", reason = FALSE, na.rm = FALSE)
}
\arguments{
\item{data}{datafame containing a vector of ORCID (XXXX-XXXX-XXXX-XXXX format)}

\item{orcid}{Column name of vector containing ORCID (default = "orcid")}

\item{reason}{Logical value to determine whether output should include reasons for NA values (default = FALSE) or vector of ORCID (TRUE).}

\item{na.rm}{Remove NA (invalid ORCID) from output (default = TRUE)}
}
\value{
Dataframe with 5 mandatory columns: orcid, full name, first name(s), last name, publication name.
}
\description{
Pull and format first name(s) and last name for a given list of orcid
}
