% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_pull_name.R
\name{orcid_pull_name}
\alias{orcid_pull_name}
\title{Pull first name(s) and last name for a given list of orcid}
\usage{
orcid_pull_name(list_orcid, initial = TRUE, initial_max = 3,
  position = "right", reason = FALSE)
}
\arguments{
\item{list_orcid}{List of orcid ids (XXXX-XXXX-XXXX-XXXX format)}

\item{initial}{Should the first / middle name(s) be converted to initial (default = TRUE)}

\item{initial_max}{Maximum number of digits (default = 3)}

\item{position}{initial to "left" or "right" of last name (default = "right")}

\item{reason}{Logical value to determine whether output should include reasons for NA values (default = FALSE) or vector of ORCID (TRUE).}
}
\value{
Dataframe with 5 mandatory columns: orcid, full name, first names, initials, and last name.
}
\description{
Pull and format first name(s) and last name for a given list of orcid
}
