% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_data.R
\name{redcap_data}
\alias{redcap_data}
\title{Export REDCap dataset and label using metadata}
\usage{
redcap_data(
  redcap_project_uri,
  redcap_project_token,
  checkbox_value = "label",
  include_original = F,
  repeat_format = "long"
)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{checkbox_value}{Determine if output checkbox variables should be unchanged from the REDCap record export ("raw") or labelled ("label"). Default = "raw".}

\item{include_original}{Logical value to determine whether original data should be provided too (default = FALSE).}

\item{repeat_format}{The format the repeating instrument data should be provided in. Options include "long" (default), "wide" (each instance a separate column), or "list" (nested instances).}
}
\value{
Three nested tibbles: (1) "exported": REDcap record export (unchanged) (2) labelled": REDcap record export with variables classified and columns labelled as specified via column_name and column_attr (3) "metadata": Cleaned metadata file for the REDCap dataset.
}
\description{
Export the REDCap dataset, and use the metadata to classify the variables and label the columns.
}
