% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_sum.R
\name{redcap_sum}
\alias{redcap_sum}
\title{Generate REDCap summary data.}
\usage{
redcap_sum(redcap_project_uri = NULL, redcap_project_token = NULL,
  use_ssl = TRUE, centre_sum = TRUE, top = 10, var_include = NULL,
  var_exclude = NULL, user_include = NULL, user_exclude = NULL,
  dag_exclude = NULL, dag_include = NULL, record_include = NULL,
  record_exclude = NULL)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{use_ssl}{Logical value whether to verify the peer's SSL certificate should be evaluated during the API pull (default=TRUE)}

\item{centre_sum}{Logical value to determine whether data access group-level summaries will be produced (Default = TRUE).}

\item{top}{When centre_sum = TRUE, defines output of the number of centres with the most records uploaded (default is top 10).}

\item{var_include}{Vector of names of variables that are desired to be specifically used to assess data completness (alternate method from using "var_exclude").}

\item{var_exclude}{Vector of names of variables that are desired to be excluded from assessment of data completness (any NA value will be counted as incomplete).}

\item{user_include}{Vector of redcap usernames that are desired to be included in the user count (note all users not assigned to a DAG will automatically be excluded).}

\item{user_exclude}{Vector of redcap usernames that are desired to be excluded from the user count (note all users not assigned to a DAG will automatically be excluded).}

\item{dag_exclude}{Vector of redcap data access group names that are desired to be excluded from the record count.}

\item{dag_include}{Vector of redcap data access group names that are desired to be included in the record count.}

\item{record_include}{Vector of redcap record_id that are desired to be included in the record count.}

\item{record_exclude}{Vector of redcap record_id that are desired to be excluded from the record count.}
}
\value{
Nested dataframes of (i) overall summary statistics for the project ("sum_overall") (ii). DAG-specific summary statistics for the project ("dag_all") (iii). DAGs with no data uploaded, but users assigned ("dag_nodata") (iv). DAGs with <100% completeness ("dag_incom") (v). The top n recruiting centres ("dag_top").
}
\description{
Used to generate high quality summary data for REDCap projects at overall, and DAG-specific level.
}
