% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_miss.R
\name{report_miss}
\alias{report_miss}
\title{Generate a missing data report for a REDCap project.}
\usage{
report_miss(redcap_project_uri, redcap_project_token, var_exclude = "")
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{var_exclude}{Vector of names of variables that are desired to be excluded from the missing data count (the default is "" e.g. none).}
}
\value{
Nested dataframe with a summary of missing data at the redcap_data_access_group level and the record level.
}
\description{
Used to generate a report of record-level + redcap_data_access_group-level missing data within a REDCap project (which accounts for branching logic in the dataframe).
}
