% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_summarise.R
\name{user_summarise}
\alias{user_summarise}
\title{Summarise the REDCap user dataframe by group (data access group)}
\usage{
user_summarise(data = NULL, redcap_project_uri = NULL,
  redcap_project_token = NULL, user_exclude = NULL, use_ssl = TRUE)
}
\arguments{
\item{data}{Dataset previously exported of users authorized for a project (5 required columns: data_access_group, username, firstname, lastname, email)}

\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{user_exclude}{Vector of usernames to be excluded e.g. those with unique rights (default = NULL).}

\item{use_ssl}{Logical value whether to verify the peer's SSL certificate should be evaluated during the API pull (default=TRUE)}
}
\value{
Dataframe summarising the user dataframe by group (data access group), number of users, and username/fullname/emails (separated by ";").
}
\description{
Group current REDCap project users by DAG (and role) to provide a summarised dataframe of users, names, and emails.
}
