% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_validate.R
\name{user_validate}
\alias{user_validate}
\title{Validates users have been allocated correctly}
\usage{
user_validate(redcap_project_uri, redcap_project_token, users.df,
  users_exception)
}
\arguments{
\item{redcap_project_uri}{URI (Uniform Resource Identifier) for the REDCap instance.}

\item{redcap_project_token}{API (Application Programming Interface) for the REDCap project.}

\item{users.df}{Dataframe containing at least 2 columns (username, data_access_group)}

\item{users_exception}{Vector of usernames to be excluded from comparison (e.g. users who should have access to the whole dataset (dag = NA) or are associated with multiple DAGs in users.df).}
}
\value{
Nested dataframe of (i) Users with NA access rights to forms (ii) Users with an unallocated DAG (iii). Users with an incorrect DAG.
}
\description{
Used to check all users have been appropriately (dag_unallocated) and correctly (dag_incorrect) allocated to a DAG, and that all user forms rights have been uploaded correctly. Note: if NA, this will default to view & edit access for all forms.
}
